requirements.txt
yfinance
finnhub-python
requests
beautifulsoup4
pandas
APScheduler
